+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - **[Fix]** Fix panic on invalid define font offset.
 - **[Fix]** Fix panic on invalid button cond key press code.
 - **[Fix]** Fix panic on unpaired morph shape record.
+- **[Fix]** Fix panic on invalid morph gradient.
 
 ## Typescript